DAT-352 documentation change, add support for maxSizeInBytes in *UnsortedStatementBatcher s

+++ executor/rxjava/src/main/java/com/datastax/dsbulk/executor/rxjava/batch/RxJavaStatementBatcher.java
@@ -127,15 +127,15 @@ public class RxJavaStatementBatcher extends StatementBatcher {
    * @param cluster The {@link Cluster} to use; cannot be {@code null}.
    * @param batchMode The batch mode to use; cannot be {@code null}.
    * @param batchType The batch type to use; cannot be {@code null}.
-   * @param maxSizeInBytes The maximum size in bytes of a batch.
    * @param maxBatchStatements - the maximum number of statements in one batch
+   * @param maxSizeInBytes The maximum size in bytes of a batch.
    */
   public RxJavaStatementBatcher(
       Cluster cluster,
       BatchMode batchMode,
       BatchStatement.Type batchType,
       int maxBatchStatements,
-      int maxSizeInBytes) {
+      long maxSizeInBytes) {
     super(cluster, batchMode, batchType, maxBatchStatements, maxSizeInBytes);
   }
